Refactor: Remove PostService and PutService files; consolidate API services
- Deleted PostService.jsx and PutService.jsx to streamline service management.
- Removed activitiesData.json to eliminate hardcoded data; replaced with dynamic data handling.
- Introduced data.js and executivesData.json for better organization of navigation links and executive member data.
- Removed unused tags.js file to clean up the codebase.
- Updated memberController.js to fix slug retrieval in deleteMember function.
- Modified memberRoutes.js to adjust the delete route to use slug as a parameter.
