+++ client/src/services/data/data.js
@@ -1,0 +1,92 @@
+export const navLinks = [
+  {
+    href: "/",
+    name: "Home",
+  },
+  {
+    href: "/about",
+    name: "About Us",
+  },
+  {
+    href: "/members",
+    name: "Members",
+  },
+  {
+    href: "/activities",
+    name: "Activities",
+  },
+  {
+    href: "/contact",
+    name: "Contact",
+  },
+];
+
+export const branches = [
+  {
+    value: "Branch - 1",
+    label: "Branch - 1",
+  },
+  {
+    value: "Branch - 2",
+    label: "Branch - 2",
+  },
+  {
+    value: "Branch - 3",
+    label: "Branch - 3",
+  },
+  {
+    value: "Main Boys",
+    label: "Main Boys",
+  },
+  {
+    value: "Main Girls",
+    label: "Main Girls",
+  },
+];
+
+export const messageTableHeader = [
+  {
+    title: "Name",
+    key: "name",
+    break: false,
+  },
+  {
+    title: "Email",
+    key: "email",
+    break: false,
+  },
+  {
+    title: "Subject",
+    key: "subject",
+    break: true,
+  },
+  {
+    title: "View",
+    key: "btn",
+    break: false,
+  },
+  {
+    title: "Action",
+    key: "btn",
+    break: false,
+  },
+];
+
+// export const tags = [
+//   {
+//     value: "Event",
+//     label: "Event",
+//   },
+//   {
+//     value: "Workshop",
+//     label: "Workshop",
+//   },
+//   {
+//     value: "Article",
+//     label: "Article",
+//   },
+//   {
+//     value: "Achievement",
+//     label: "Achievement",
+//   },
+// ];
