--- client/src/services/data/tags.js
@@ -1,20 +1,0 @@
-const tags = [
-  {
-    value: "Event",
-    label: "Event",
-  },
-  {
-    value: "Workshop",
-    label: "Workshop",
-  },
-  {
-    value: "Article",
-    label: "Article",
-  },
-  {
-    value: "Achievement",
-    label: "Achievement",
-  },
-];
-
-export default tags;
